Use picocss, avoid forms for simple POSTs
I wanted a minimal css framework. Picocss seems nice, and is largely
just some nice syling over core html elements.

As the prominant feature of this project is a filestorage microservice,
I didn't like having to send multipart form data for all POSTs, instead
preferring json. Additionally, FastAPI isn't as rich around form data as
it is json or url encoded requests. Might make it easier to consume as
an API. Doing this in htmx is a bit funky, but not terrible.